Removed styles management.
The amount of styles is too low to justify using SCSS. The styles will be now managed through a monolithic already minified "styles.js" file, which was moved to "src/styles.js".
The "src/dev.js" file, which content was solely the inclusion of the "src/index.js" file and the inclusion of the styles file, was removed. The "webpack.config.js" file was accordingly adjusted.