Removed copying of the production files to the "dist" directory.
The production files—"index.js", "README.md", "package.json", and "LICENSE"—will be now zipped directly from the project's directory.
Thanks to this the automatic copying of these files will be no longer needed, so it implies one package less—"copy-webpack-plugin"—which was removed.
The ignore files were removed to respect the removal of the directory.