Removed unneeded lines from the ".npmignore" file and ordered its content.
The removed lines are the "tests" directory and the "gulpfile.babel.js" file—left after the removal of tests and styles respectively.
The content was order alphabetically.
Additionally, the slashes from the beginnings of directories were removed and added to the end of them so as to respect how does it look like in a UNIX system.