Removed the "src" directory and moved its content to the project's directory.
There are too few files in it to justify making a separate directory, especially if its content is the core, the most often edited part of the project. So the access to the highly edited files will be faster.
The "src/index.js" file was removed to "i.js" file as the "index.js" file is the production file destined to the npm, Inc.'s registry. The webpack configuration files were updated to respect the change.