The "index.js" file is now copied after running the production script.
There have to be two production files: one—"index.js"—for the npm, Inc.'s registry and another—"fslightbox.js"—for the production archive. The "fslightbox.js" file is copied from the "index.js" file through "&& cp index.js fslightbox.js". It's working solely in a UNIX system, but that's fine for me. In my opinion it's better than adding a whole package, such as "copy-webpack-plugin", when chances that the command will have to be run from a non-UNIX system are low.