Trimmed the demo.
It's now displaying only an image and an invalid source. HTML videos, YouTube videos, and custom sources are too rarely tested to make justifiable connecting to external servers on every open of the test lightbox.

The "demo.css" file was removed as a Vimeo video is no longer displayed.

The "1.jpg" image was replaced with an image of smaller resolution.