Squeezed and shortened the names of the smw and saw wrappers, and moved the loader to the saw wrapper.
The loader was previously in the smw wrapper. It's way of creation was changed, now it's not created with an HTML string, but with DOM manipulation functions as the "document.createElement" function. The "fslightboxlc" class was added to the inner loader elements—previously the styles of these elements were managed through the <div> tag selector. The set loader class was shortened to "fslightboxl", but the names of updated classes in the JavaScript were not yet applied to CSS.