Removed the SourceMainWrapperTransform object and moved its object to the smw element.
This change was done because I deem those the smw DOM element and the SourceMainWrapperTransformer as too close to each other. The SourceMainWrapperTransformer is a small object, which is doing only simply DOM manipulations, so why not merge it with the object it's manipulating?

The changes were not yet reflected in the places where the previous object was used.